% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BioprodNamibianHake}
\alias{BioprodNamibianHake}
\title{Biomass Production for Namibian Hake over 25 years}
\format{
A named list:
\describe{
  \item{N}{The number of years in the data set}
  \item{Year}{The years of the data}
  \item{C}{Number of catches in thousand tons}
  \item{I}{Abundance indices based on CPUE in tons per standadized trawler hours}
}
}
\source{
McAllister, M. K., & Kirkwood, G. P. (1998). Bayesian stock assessment: a review and example application using the logistic model. ICES Journal of Marine Science, 55(6), 1031-1060.
}
\usage{
BioprodNamibianHake
}
\description{
A dataset containing the biomass and abundance indices from
1964 to 1988.
}
\keyword{datasets}
