% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CmrOir}
\alias{CmrOir}
\title{Catch-mark-recapture data of adult salmon on the Oir river}
\format{
A named list:
\describe{
  \item{N}{The number of years in the data set}
  \item{y1}{The number of trapped fish}
  \item{y2}{Number of one sea-winter (1SW) fish removed from the population}
  \item{y3}{NUmber of two sea-winter (2SW) fish removed from the population}
  \item{y4}{Number of tagged and released fish}
  \item{y5}{Number of marked-recapture fish}
  \item{y6}{Number of unmarked-recapture fish}
}
}
\source{
Rivot, E., & Prévost, E. (2002). Hierarchical Bayesian analysis of capture mark recapture data. Canadian Journal of Fisheries and Aquatic Sciences, 59(11), 1768-1784.
}
\usage{
CmrOir
}
\description{
A dataset containing CMR data from 1984 to 2000
}
\keyword{datasets}
