% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_longer}
\alias{extract_longer}
\title{Extract object from various MCMC sampler into a long data frame}
\usage{
extract_longer(fit)
}
\arguments{
\item{fit}{An object of type \code{stanfit} or \code{mcmc.list}}
}
\value{
A tibble object of long format
}
\description{
The dataset as the form [iteration, chain, parameter, value]
}
