% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_wider}
\alias{extract_wider}
\title{Extract object from various MCMC sampler into a wide data frame}
\usage{
extract_wider(fit)
}
\arguments{
\item{fit}{An object of type \code{stanfit} or \code{mcmc.list}}
}
\value{
A tibble object of wide format
}
\description{
The dataset as the form [iteration, chain, parameters]
}
