% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SalmonLifeCycle}
\alias{SalmonLifeCycle}
\title{Life cycle data of Atlantic salmon}
\format{
A named list:
\describe{
  \item{n}{The number of years}
  \item{surf}{The wetted surface of the system}
  \item{fec1}{Fecunity (number of eggs / female) of 1 Sea Winter (1SW) fish}
  \item{fec2}{Fecunity (number of eggs / female) of 2 Sea Winter (2SW) fish}
  \item{Year}{The years of the data}
  \item{c_Sp}{Number of of spawners trapped during upstream migration
  (\eqn{y_1^{Sp}})}
  \item{x_Sp1}{Number of 1SW Spawners retrieved fron the
  population (\eqn{y_2^{Sp}})}
  \item{x_Sp2}{Number of 2SW Spawners retrieved fron the
  population (\eqn{y_3^{Sp}})}
  \item{mad}{Marked and released Spawners (\eqn{y_4^{Sp}})}
  \item{rmad}{Marked and recaptured Spawners (\eqn{y_5^{Sp}})}
  \item{r_unm}{Unmarked and recaptured Spawners (\eqn{y_6^{Sp}})}
  \item{sample_Sp_age}{Adults examined for ageing (\eqn{y_7^{Sp}})}
  \item{sample_Sp11}{Number of 1SW issued from 1+ Smolts among sample_Sp_age
  (\eqn{y_8^{Sp}})}
  \item{sample_Sp21}{Number of 1SW issued from 2+ Smolts among sample_Sp_age
  (\eqn{y_9^{Sp}})}
  \item{sample_Sp12}{Number of 2SW issued from 1+ Smolts among sample_Sp_age
  (\eqn{y_{10}^{Sp}})}
  \item{sample_Sp22}{Number of 2SW issued from 2+ Smolts among sample_Sp_age
  (\eqn{y_{11}^{Sp}})}
  \item{sample_Sp1_sex}{Number of 1SW fish examined for sex identification
  (\eqn{y_{12}^{Sp}})}
  \item{sample_Sp2_sex}{Number of 2SW fish examined for sex identification
  (\eqn{y_{13}^{Sp}})}
  \item{sample_Sp1f}{Number of 1SW fish identified as female among
  sample_Sp1_sex (\eqn{y_{14}^{Sp}})}
  \item{sample_Sp2f}{Number of 1SW fish identified as female among
  sample_Sp2_sex (\eqn{y_{15}^{Sp}})}
  \item{c_Sm}{Number of smolts  caught in the downstream trapping facility
 during the migration time(\eqn{y_{1}^{Sm}})}
  \item{m_Sm}{Number of smolts  marked and released (\eqn{y_{2}^{Sm}})}
  \item{m_Sm}{Number of smolts  marked-released and recaptured (\eqn{y_{3}^{Sm}})}
  \item{sample_Sm_age}{Number of smolts  examined for river-age (\eqn{y_{4}^{Sm}})}
  \item{sample_Sm1}{Number of 1+ smolts  among sample_Sm_age (\eqn{y_{5}^{Sm}})}
}
}
\source{
Rivot, E., Prévost, E., Cuzol, A., Parent, E. & Baglinière, J. L.
(2004). A Bayesian state-space modeling framework for fitting a salmon
stage-structured population dynamic model to multiple time series of
field data. Ecological Modelling, 179, 463-485.
}
\usage{
SalmonLifeCycle
}
\description{
A dataset resulting from a comprehensive survey of the salmon
population of the Oir river between 1984 and 2001. The dataset corresponds
essentially to Catch Mark Recapture experiments.
}
\keyword{datasets}
