% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SRSalmon}
\alias{SRSalmon}
\title{Stock recruitment data of A. salmon over 13 rivers}
\format{
A named list:
\describe{
  \item{n_riv}{The number of rivers}
  \item{name_riv}{The name of the rivers}
  \item{country_riv}{The country of the rivers}
  \item{area}{The riverine wetted area accessible to salmon in
  squared meters}
  \item{n_obs}{The number of observations for each river}
  \item{n}{The total number of observations}
  \item{riv}{The river membership of each observation}
  \item{S}{Stock data of length n}
  \item{R}{Recruitment data of length n}
  \item{lat}{Latitude of each river site}
  \item{n_pred}{Number of predictions}
  \item{lat_pred}{Latitude used for predictions}
}
}
\source{
Prévost, E., Parent, E., Crozier, W., Davidson, I., Dumas, J.,
Gudbergsson, G., ... & Sættem, L. M. (2003). Setting biological reference
points for Atlantic salmon stocks: transfer of information from data-rich to
sparse-data situations by Bayesian hierarchical modelling.
ICES Journal of Marine Science, 60(6), 1177-1193.
}
\usage{
SRSalmon
}
\description{
A dataset containing the stock and recruitment of A. salmon over 13 european
rivers. As well as the latitude and name of the different site and new
latitude to perform predictions.
}
\keyword{datasets}
