% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SucRemNivelle}
\alias{SucRemNivelle}
\title{Three passes successive removal data of A. salmon in the Nivelle river}
\format{
A named list:
\describe{
  \item{I}{The number of sites (sites x years)}
  \item{C1}{The number of captures of the first pass for each site}
  \item{C2}{The number of captures of the second pass for each site}
  \item{C3}{The number of captures of the third pass for each site}
  \item{S}{The area in squared meter that was fished for each site}
  \item{AI}{The abundance index of each site}
  \item{I_pred}{The number of sites (for prediction)}
  \item{AI_pred}{Abundance indices (for prediction)}
}
}
\source{
Brun, M., Abraham, C., Jarry, M., Dumas, J., Lange, F., & Prévost, E.
 (2011). Estimating an homogeneous series of a population abundance indicator
  despite changes in data collection procedure: A hierarchical Bayesian
  modelling approach. Ecological Modelling, 222(5), 1069-1079.
}
\usage{
SucRemNivelle
}
\description{
A dataset containing the successive removal data electofishing of
salmon
juveniles over 3 passes on 3 years (2003--2005) and 11 sites. Contains
missing data.
}
\keyword{datasets}
