% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SucRemOir}
\alias{SucRemOir}
\title{Successive removal data of A. salmon in the Oir river}
\format{
A named list:
\describe{
  \item{ny}{The number of years}
  \item{smolts}{Number of smolts for cohort of 1986--2003.
  Cohorts match with juveniles. The fifth entry (smolts = 169) corresponds to
  year 1990 for which no 0+ juveniles data is available.}
  \item{n}{Number of observations}
  \item{Year}{The year of the data. 1990 is missing}
  \item{Y}{Index of the years}
  \item{nsy}{The cumulative number of sites per year}
  \item{H}{The habitat type. 1: Rapid/riffle; 2: Run}
  \item{C1}{Number of captures at the first pass}
  \item{C2}{Number of captures at the second pass}
  \item{S}{The surface of each site in squared meter}
  \item{Dat}{Number of predictions}
  \item{n_sites_ex}{The number of sites for extrapolation}
  \item{s_ex}{Mean surface per site per habitat type (for extrapolation)}
}
}
\source{
Rivot, E., Prévost, E., Cuzol, A., Baglinière, J. L., & Parent, E.
(2008). Hierarchical Bayesian modelling with habitat and time covariates for
estimating riverine fish population size by successive removal method.
Canadian Journal of Fisheries and Aquatic Sciences, 65(1), 117-133.
}
\usage{
SucRemOir
}
\description{
A dataset containing successive removal data with two passes of
Atlantic salmon over 20 years and between 7 and 10 sampling sites per year.
The data also contains the surface area of the sampling sites.
}
\keyword{datasets}
