% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SucRemNivelle}
\alias{SucRemNivelle}
\title{Successive removal data of A. salmon in the Nivelle river}
\format{
A named list:
\describe{
  \item{n_riv}{The number of rivers}
  \item{name_riv}{The name of the rivers}
  \item{country_riv}{The country of the rivers}
  \item{area}{The riverine wetted area accessible to salmon in
  squared meters}
  \item{n_obs}{The number of observations for each river}
  \item{n}{The total number of observations}
  \item{riv}{The river membership of each observation}
  \item{S}{Stock data of length n}
  \item{R}{Recruitment data of length n}
  \item{lat}{Latitude of each river site}
  \item{n_pred}{Number of predictions}
  \item{lat_pred}{Latitude used for predictions}
}
}
\source{
McAllister, M. K., & Kirkwood, G. P. (1998). Bayesian stock assessment: a review and example application using the logistic model. ICES Journal of Marine Science, 55(6), 1031-1060.
}
\usage{
SucRemNivelle
}
\description{
A dataset containing the stock and recruitment of A. salmon over 13 european
rivers. As well as the latitude and name of the different site and new
latitude to perform predictions.
}
\keyword{datasets}
